% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDP_mike.R
\name{calcGDP_mike}
\alias{calcGDP_mike}
\title{Calculate GDP}
\usage{
calcGDP_mike(population, gdpPercapita)
}
\arguments{
\item{population}{the population value}

\item{gdpPercapita}{the gdp value}
}
\description{
Calculate the Gross Domestic Product
}
\details{
Takes population and gdp percapita and multiplies them together
}
\examples{
calcGDP_mike(10, 2.3)
}

